coastsnap-web/backend/requirements.txt
# Requires Python >= 3.8
Flask==2.3.2
flask-cors==4.0.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lightglue @ git+https://github.com/cvg/LightGlue.git@edb2b838efb2ecfe3f88097c5fad9887d95aedad
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.0
pandas==2.1.4
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.13.1
scikit-image==0.22.0
six==1.16.0
tbb==2021.12.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
openpyxl==3.1.2
